+++ braintree/bank_account_instant_verification_jwt.py
@@ -1,0 +1,13 @@
+from braintree.attribute_getter import AttributeGetter
+
+class BankAccountInstantVerificationJwt(AttributeGetter):
+    def __init__(self, jwt):
+        self.jwt = jwt
+
+    @property
+    def jwt(self):
+        return self._jwt
+
+    @jwt.setter
+    def jwt(self, value):
+        self._jwt = value