+++ braintree/braintree_gateway.py
@@ -1,6 +1,7 @@
 from braintree.add_on_gateway import AddOnGateway
 from braintree.address_gateway import AddressGateway
 from braintree.apple_pay_gateway import ApplePayGateway
+from braintree.bank_account_instant_verification_gateway import BankAccountInstantVerificationGateway
 from braintree.client_token_gateway import ClientTokenGateway
 from braintree.configuration import Configuration
 from braintree.credit_card_gateway import CreditCardGateway
@@ -47,6 +48,7 @@ class BraintreeGateway(object):
         self.add_on = AddOnGateway(self)
         self.address = AddressGateway(self)
         self.apple_pay = ApplePayGateway(self)
+        self.bank_account_instant_verification = BankAccountInstantVerificationGateway(self)
         self.client_token = ClientTokenGateway(self)
         self.credit_card = CreditCardGateway(self)
         self.customer = CustomerGateway(self)