+++ braintree/sender.py
@@ -1,0 +1,5 @@
+from braintree.attribute_getter import AttributeGetter
+
+class Sender(AttributeGetter):
+    def __init__(self, attributes):
+        AttributeGetter.__init__(self, attributes)