+++ braintree/us_bank_account_verification.py
@@ -31,12 +31,14 @@ class UsBankAccountVerification(AttributeGetter):
 
         * braintree.UsBankAccountVerification.VerificationMethod.NetworkCheck
         * braintree.UsBankAccountVerification.VerificationMethod.IndependentCheck
+        * braintree.UsBankAccountVerification.VerificationMethod.InstantVerificationAccountValidation
         * braintree.UsBankAccountVerification.VerificationMethod.TokenizedCheck
         * braintree.UsBankAccountVerification.VerificationMethod.MicroTransfers
         """
 
         NetworkCheck = "network_check"
         IndependentCheck = "independent_check"
+        InstantVerificationAccountValidation = "instant_verification_account_validation"
         TokenizedCheck = "tokenized_check"
         MicroTransfers = "micro_transfers"